move prepare_ead_1 and prepare_ead_3 to attestation library, clean the printf
